+++ src/utils/index.js
@@ -1,5 +1,6 @@
 export { default as clean } from "./clean";
 export { default as evaluate } from "./evaluate";
+export { renderers } from "./markdown-renderers";
 export { default as pluralize } from "./pluralize";
 export { default as shorten } from "./shorten";
 export { default as unique } from "./unique";