+++ src/utils/markdown-renderers.jsx
@@ -1,23 +1,18 @@
-import React from "react";
-
-import ReactMarkdown from "react-markdown";
 import { Link } from "react-router-dom";
 
 import Typography from "@material-ui/core/Typography";
 
-const renderers = {
+export const renderers = {
   emphasis: ({ children }) => (
     <Typography component="em" variant="inherit">
       {children}
     </Typography>
   ),
   heading: ({ children, level }) => (
-    <Typography variant={`h${[3, 3, 4, 5, 6, 6, 6][level]}`}>
-      {children}
-    </Typography>
+    <Typography variant={`h${level}`}>{children}</Typography>
   ),
   link: ({ children, href, title }) =>
-    href.match(/^(https?)\/\//) ? (
+    href.match(/^https?:\/\//) ? (
       <a href={href} title={title} target="_blank" rel="noopener noreferrer">
         {children}
       </a>
@@ -43,14 +38,3 @@ const renderers = {
     </Typography>
   ),
 };
-
-/**
- * `<Markdown source={"# Markdown"}/>`
- * @constructor
- */
-export default function Markdown(props) {
-  console.warn(
-    "`Markdown` is deprecated since v1.29.0 and will be removed in v2.0.0, use `react-markdown` instead. A list of renderers is available in `mastro-elfo-mui/utils/markdown-renderers`"
-  );
-  return <ReactMarkdown {...props} renderers={renderers} />;
-}